software/gcs/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pymavlink